+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: tlemos-m <tlemos-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/14 15:53:27 by tlemos-m          #+#    #+#             */
-/*   Updated: 2022/11/18 13:59:19 by tlemos-m         ###   ########.fr       */
+/*   Updated: 2022/11/23 10:46:52 by tlemos-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int		ft_printf(const char *s, ...);
 void	ft_putchar(char c);
 int		ft_putstr(char *s);
 int		ft_putnbr_base(long long n, char *base, int b);
-int		ft_putptr_base(unsigned long long n, char *base, int j);
+int		ft_putptr_base(unsigned long n, char *base, int j);
 size_t	ft_strlen(char *s);
 
 #endif
